Core/Guilds: First step in correctly implementing cataclysm guild system
Core/currency: Changed money to 'uint64' to satisfy client
Repo Codestlye changes:
pCreature->creature
pPlayer->player
plr->player
pl->player (and cleaned up a few pointers)
pGuild->guild

TODO:
Re-Check Opcodes and handlers, still getting not allowed opcode spam
Finish missing pieces to guild system, seem the system works but is not initializing on the players side
also track down wowerror on guild disban.

+++ src/server/game/Mails/Mail.h
@@ -184,8 +184,8 @@ struct Mail
     std::vector<uint32> removedItems;
     time_t expire_time;
     time_t deliver_time;
-    uint32 money;
-    uint32 COD;
+    uint64 money;
+    uint64 COD;
     uint32 checked;
     MailState state;
