+++ src/server/game/Server/Protocol/Handlers/PetitionsHandler.cpp
@@ -103,7 +103,7 @@ void WorldSession::HandlePetitionBuyOpcode(WorldPacket & recv_data)
         GetPlayer()->RemoveAurasByType(SPELL_AURA_FEIGN_DEATH);
 
     uint32 charterid = 0;
-    uint32 cost = 0;
+    uint64 cost = 0;
     uint32 type = 0;
     if (creature->isTabardDesigner())
     {
